refactor: Controller 요청 값에 대한 예외 처리 추가

+++ src/main/java/com/thekey/stylekeyserver/coordinatelook/service/CoordinateLookAdminServiceImpl.java
@@ -1,10 +1,8 @@
 package com.thekey.stylekeyserver.coordinatelook.service;
 
 import static com.thekey.stylekeyserver.common.exception.ErrorCode.COORDINATE_LOOK_NOT_FOUND;
-import static com.thekey.stylekeyserver.common.exception.ErrorCode.ITEM_NOT_FOUND;
 
 import com.thekey.stylekeyserver.common.exception.ApiException;
-import com.thekey.stylekeyserver.common.exception.ApiResponse;
 import com.thekey.stylekeyserver.common.exception.ErrorCode;
 import com.thekey.stylekeyserver.coordinatelook.domain.CoordinateLook;
 import com.thekey.stylekeyserver.coordinatelook.dto.request.CoordinateLookRequest;
@@ -19,14 +17,10 @@ import com.thekey.stylekeyserver.item.service.ItemAdminService;
 import com.thekey.stylekeyserver.common.s3.service.S3Service;
 import com.thekey.stylekeyserver.stylepoint.domain.StylePoint;
 import com.thekey.stylekeyserver.stylepoint.service.StylePointAdminService;
-import java.io.IOException;
-import java.io.UnsupportedEncodingException;
-import java.net.MalformedURLException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.NoSuchElementException;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
@@ -56,7 +50,13 @@ public class CoordinateLookAdminServiceImpl implements CoordinateLookAdminServic
     @Transactional
     public CoordinateLook create(CoordinateLookRequest requestDto,
                                  MultipartFile coordinateLookImageFile,
-                                 List<MultipartFile> itemImageFiles) throws Exception {
+                                 List<MultipartFile> itemImageFiles) {
+
+        if(requestDto == null) {
+            throw new ApiException(ErrorCode.ERROR_BAD_REQUEST);
+        }
+
+        validateImageFiles(coordinateLookImageFile, itemImageFiles);
 
         StylePoint stylePoint = stylePointAdminService.findById(requestDto.getStylePointId());
         CoordinateLook coordinateLook = requestDto.toEntity(stylePoint);
@@ -74,10 +74,6 @@ public class CoordinateLookAdminServiceImpl implements CoordinateLookAdminServic
         for (ItemRequest itemDto : itemRequests) {
             Long itemId = itemDto.getId();
             MultipartFile itemImageFile = itemImageFilesMap.get(itemId);
-
-            if (itemImageFile == null) {
-                throw new ApiException(ErrorCode.INVALID_IMAGE_FORMAT);
-            }
 
             Item item = itemAdminService.create(itemDto, itemImageFile);
             items.add(item);
@@ -117,30 +113,32 @@ public class CoordinateLookAdminServiceImpl implements CoordinateLookAdminServic
 
     @Override
     @Transactional
-    public CoordinateLook update(Long id, CoordinateLookRequest requestDto,
-                                 MultipartFile coordinateLookImageFile) throws IOException {
+    public CoordinateLook update(Long id, CoordinateLookRequest requestDto, MultipartFile coordinateLookImageFile) {
+
+        if(requestDto == null) {
+            throw new ApiException(ErrorCode.ERROR_BAD_REQUEST);
+        }
 
         CoordinateLook coordinateLook = coordinateLookRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(
                         COORDINATE_LOOK_NOT_FOUND.getMessage() + id));
         StylePoint stylePoint = stylePointAdminService.findById(requestDto.getStylePointId());
 
-        if (!coordinateLookImageFile.isEmpty()) {
+        // 수정 할 이미지가 요청값에 포함 되어있을 때만 기존 이미지에서 수정할 이미지로 변경
+        if (coordinateLookImageFile != null && !coordinateLookImageFile.isEmpty()) {
             Image coordinateLookOldImage = coordinateLook.getImage();
+            coordinateLookOldImage.setUnused();
+            imageRepository.save(coordinateLookOldImage);
+            imageService.deleteUnusedImages();
 
-            if (coordinateLookOldImage != null) {
-                coordinateLookOldImage.setUnused();
-                imageRepository.save(coordinateLookOldImage);
-                imageService.deleteUnusedImages();
+            Image newImage = s3Service.uploadFile(coordinateLookImageFile, Type.COORDINATE_LOOK);
+            imageRepository.save(newImage);
 
-                Image newImage = s3Service.uploadFile(coordinateLookImageFile, Type.COORDINATE_LOOK);
-                imageRepository.save(newImage);
-
-                coordinateLook.setImage(newImage);
-                coordinateLookRepository.save(coordinateLook);
-            }
+            coordinateLook.setImage(newImage);
+            coordinateLookRepository.save(coordinateLook);
         }
 
+        // 수정 할 이미지가 없다면 기본 정보만 변경
         coordinateLook.update(
                 requestDto.getTitle(),
                 stylePoint);
@@ -150,18 +148,23 @@ public class CoordinateLookAdminServiceImpl implements CoordinateLookAdminServic
     @Override
     @Transactional
     public CoordinateLook updateItem(Long coordinateLookId, Long itemId, ItemRequest requestDto,
-                                     MultipartFile itemImageFile) throws IOException {
+                                     MultipartFile itemImageFile) {
+
+        if(requestDto == null) {
+            throw new ApiException(ErrorCode.ERROR_BAD_REQUEST);
+        }
 
         CoordinateLook coordinateLook = coordinateLookRepository.findById(coordinateLookId)
                 .orElseThrow(() -> new EntityNotFoundException(
                         COORDINATE_LOOK_NOT_FOUND.getMessage() + coordinateLookId));
+
         itemAdminService.update(coordinateLookId, itemId, requestDto, itemImageFile);
         return coordinateLook;
     }
 
     @Override
     @Transactional
-    public void delete(Long id) throws MalformedURLException, UnsupportedEncodingException {
+    public void delete(Long id) {
         CoordinateLook coordinateLook = coordinateLookRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(
                         COORDINATE_LOOK_NOT_FOUND.getMessage() + id));
@@ -194,24 +197,25 @@ public class CoordinateLookAdminServiceImpl implements CoordinateLookAdminServic
 
         List<Item> items = coordinateLook.getItems();
 
-        try {
-            Item itemToDelete = items.stream()
-                    .filter(item -> item.getId().equals(itemId))
-                    .findFirst()
-                    .orElseThrow();
+        Item itemToDelete = items.stream()
+                .filter(item -> item.getId().equals(itemId))
+                .findFirst()
+                .orElseThrow();
 
-            itemAdminService.delete(itemToDelete.getId());
-            items.remove(itemToDelete);
+        itemAdminService.delete(itemToDelete.getId());
+        items.remove(itemToDelete);
 
-            // 코디룩에 속한 아이템이 없을 경우, 코디룩도 함께 삭제한다.
-            if (coordinateLook.getItems().isEmpty()) {
-                coordinateLookRepository.delete(coordinateLook);
-            }
-        } catch (NoSuchElementException | MalformedURLException | UnsupportedEncodingException e) {
-            throw new EntityNotFoundException(ITEM_NOT_FOUND.getMessage() + itemId);
+        // 코디룩에 속한 아이템이 없을 경우, 코디룩도 함께 삭제한다.
+        if (coordinateLook.getItems().isEmpty()) {
+            coordinateLookRepository.delete(coordinateLook);
         }
     }
 
+    private void validateImageFiles(MultipartFile coordinateLookImageFile, List<MultipartFile> itemImageFiles) {
+        if (coordinateLookImageFile == null || coordinateLookImageFile.isEmpty() || itemImageFiles == null || itemImageFiles.isEmpty()) {
+            throw new ApiException(ErrorCode.FILE_NOT_FOUND);
+        }
+    }
 
     private Long getItemIdFromFileName(MultipartFile fileName) {
         String filenameWithoutExtension = fileName.getOriginalFilename()
